esdm_es: fix method signature

+++ addon/linux_esdm_es/esdm_health.h
@@ -142,7 +142,7 @@ enum esdm_health_res {
 bool esdm_sp80090b_startup_complete_es(enum esdm_internal_es es);
 bool esdm_sp80090b_compliant(enum esdm_internal_es es);
 
-enum esdm_health_res esdm_health_test(u32 now_time, enum esdm_internal_es es);
+enum esdm_health_res esdm_health_test(u64 now_time, enum esdm_internal_es es);
 void esdm_health_disable(void);
 
 #endif /* _ESDM_HEALTH_H */
